Turn whitelist activity into a full-fledged activity
Instead of having this tiny dialog like thing, use a full
activity, and add app icons to the list. While these make
the list a lot longer, they should improve usability quite
a lot as they provide a visual clue as to the identity of
the apps.

Also turn the checkbox into a switch. That just looks better
in this context.